+++ src/profile/scripts/app/controllers/dialogs/warningDialogController.js
@@ -1,0 +1,23 @@
+/**
+ * Created by szhitenev on 08.06.2016.
+ */
+(function(){
+
+    'use strict';
+
+    module.exports = function($scope, $mdDialog, warning){
+
+        var vm = this;
+
+        vm.warning = warning;
+
+        vm.cancel = function () {
+            $mdDialog.cancel();
+        };
+
+        vm.agree = function () {
+            $mdDialog.hide({status: 'agree'});
+        };
+    }
+
+}());